split constraint types up into separate files, allow user to import directly
